+++ toggle-control-raw.js
@@ -6,7 +6,7 @@ class ToggleControl extends GroundControl {
         ToggleControl
       );
     }
-  }
+  };
 
   toggles = [];
 
@@ -17,23 +17,23 @@ class ToggleControl extends GroundControl {
         target === btn ? 'true' : 'false'
       );
     });
-  }
+  };
 
   get pressed() {
     return this.toggles.find(
       (btn) => btn.getAttribute('aria-pressed') === 'true'
     );
-  }
+  };
 
   get pressedValue() {
     return this.pressed?.dataset.value;
-  }
+  };
 
   constructor() {
     super();
     GroundControl.blockDisplay(this);
     this.addEventListener('click', this.onTogglePress);
-  }
+  };
 
   connectedCallback() {
     super.connectedCallback();
@@ -60,18 +60,18 @@ class ToggleControl extends GroundControl {
     this.initialValue = this.pressedValue || this.dataset.off;
 
     this.value = this.storedValue || this.initialValue;
-  }
+  };
 
   onTogglePress = (event) => {
     this.doPress(event.target);
     this.value = this.pressedValue;
-  }
+  };
 
   onValueChange = () => {
     if (this.usedValue !== this.pressedValue) {
       this.pressed = this.#toggleFromValue();
     }
-  }
+  };
 
   doPress = (target) => {
     if (!this.toggles.includes(target)) return;
@@ -84,7 +84,7 @@ class ToggleControl extends GroundControl {
     } else {
       this.pressed = target;
     }
-  }
+  };
 
   #toggleFromValue = () => this.toggles.find(
     (btn) => btn.dataset.value === this.usedValue